aprgcommon refactoring

+++ AprgCommon/AprgCommon/src/Common/Math/Helpers/FractionHelpers.cpp
@@ -74,11 +74,10 @@ void changeFractionToSimplestForm(NumberType1 & numerator, NumberType2 & denomin
 }
 
 // instantiation should fit on long long int
-template void changeFractionToSimplestForm<int, unsigned int, long long int>(int & numerator, unsigned int & denominator);
+template void changeFractionToSimplestForm<int32_t, uint32_t, int64_t>(int32_t & numerator, uint32_t & denominator);
 
 FractionDetails getBestFractionDetailsForDoubleValue(
-        double const doubleValue)
-{
+        double const doubleValue){
     constexpr double tolerance(1E-12);
     FractionDetails result{1, 0, 1};
