+++ AprgCommon/AprgCommon/tst/CommonTests/Math/Helpers/PrecisionHelpers_unit.cpp
@@ -80,21 +80,19 @@ TEST(PrecisionHelpersTest, IsValueWithinShortIntegerLimitsWorks)
 TEST(PrecisionHelpersTest, IsValueWithinLongIntegerLimitsWorks)
 {
     EXPECT_FALSE(isValueWithinLimits<long int>(static_cast<double>(LONG_MIN)-100000));
-    EXPECT_TRUE(isValueWithinLimits<long int>(LONG_MIN));
+    EXPECT_TRUE(isValueWithinLimits<long int>(static_cast<double>(LONG_MIN)));
     EXPECT_TRUE(isValueWithinLimits<long int>(1000));
-    EXPECT_TRUE(isValueWithinLimits<long int>(LONG_MAX));
+    EXPECT_TRUE(isValueWithinLimits<long int>(static_cast<double>(LONG_MAX)));
     EXPECT_FALSE(isValueWithinLimits<long int>(static_cast<double>(LONG_MAX)+100000));
 }
-
 TEST(PrecisionHelpersTest, IsValueWithinUnsignedLongIntegerLimitsWorks)
 {
     EXPECT_FALSE(isValueWithinLimits<unsigned long int>(static_cast<double>(0)-100000));
     EXPECT_TRUE(isValueWithinLimits<unsigned long int>(0));
     EXPECT_TRUE(isValueWithinLimits<unsigned long int>(1000));
-    EXPECT_TRUE(isValueWithinLimits<unsigned long int>(ULONG_MAX));
+    EXPECT_TRUE(isValueWithinLimits<unsigned long int>(static_cast<double>(ULONG_MAX)));
     EXPECT_FALSE(isValueWithinLimits<unsigned long int>(static_cast<double>(ULONG_MAX)+100000));
 }
-
 TEST(PrecisionHelpersTest, IsValueWithinLongLongIntegerLimitsWorks)
 {
     EXPECT_FALSE(isValueWithinLimits<long long int>(static_cast<double>(LLONG_MIN)*2));
@@ -143,21 +141,19 @@ TEST(PrecisionHelpersTest, IsValueBeyondShortIntegerLimitsWorks)
 TEST(PrecisionHelpersTest, IsValueBeyondLongIntegerLimitsWorks)
 {
     EXPECT_TRUE(isValueBeyondLimits<long int>(static_cast<double>(LONG_MIN)-100000));
-    EXPECT_FALSE(isValueBeyondLimits<long int>(LONG_MIN));
+    EXPECT_FALSE(isValueBeyondLimits<long int>(static_cast<double>(LONG_MIN)));
     EXPECT_FALSE(isValueBeyondLimits<long int>(1000));
-    EXPECT_FALSE(isValueBeyondLimits<long int>(LONG_MAX));
+    EXPECT_FALSE(isValueBeyondLimits<long int>(static_cast<double>(LONG_MAX)));
     EXPECT_TRUE(isValueBeyondLimits<long int>(static_cast<double>(LONG_MAX)+100000));
 }
-
 TEST(PrecisionHelpersTest, IsValueBeyondUnsignedLongIntegerLimitsWorks)
 {
     EXPECT_TRUE(isValueBeyondLimits<unsigned long int>(static_cast<double>(0)-100000));
     EXPECT_FALSE(isValueBeyondLimits<unsigned long int>(0));
     EXPECT_FALSE(isValueBeyondLimits<unsigned long int>(1000));
-    EXPECT_FALSE(isValueBeyondLimits<unsigned long int>(ULONG_MAX));
+    EXPECT_FALSE(isValueBeyondLimits<unsigned long int>(static_cast<double>(ULONG_MAX)));
     EXPECT_TRUE(isValueBeyondLimits<unsigned long int>(static_cast<double>(ULONG_MAX)+100000));
 }
-
 TEST(PrecisionHelpersTest, IsValueBeyondLongLongIntegerLimitsWorks)
 {
     EXPECT_TRUE(isValueBeyondLimits<long long int>(static_cast<double>(LLONG_MIN)*2));
